+++ src/components/pages/Gamemode/Gamemode.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React, { useEffect } from 'react';
 // import history from './history';
 import { connect } from 'react-redux';
 import { tasks } from '../../../state/actions';
@@ -8,57 +8,43 @@ import GamemodeButton from './GamemodeButton';
 
 import { render } from 'react-dom';
 
-const Gamemode = () => {
+const Gamemode = ({ ...props }) => {
   const { push, location } = useHistory();
-  const [sP, setsP] = useState(false);
+  // const [sP, setsP] = useState(false);
   //   history.push('/gamemode/single');
-  const ff = () => {
-    if (location.pathname === '/gamemode/single' && sP === true) {
-      push('/gamemode');
-      setsP(false);
-    } else if (location.pathname === '/gamemode' && sP === false) {
-      push('/gamemode/single');
-      setsP(true);
-    }
-  };
 
   //   useEffect(() => {
 
   //     ff();
   //   }, []);
 
-  const singled = () => {
-    if (location.pathname === '/gamemode/single' && sP === true) {
-      push('/gamemode');
-      setsP(false);
-      console.log('nal', this.state.hasRead);
-    } else if (location.pathname === '/gamemode' && sP === false) {
-      push('/gamemode/single');
-      setsP(true);
-    }
-  };
+  // const singled = () => {
+  //   if (location.pathname === '/gamemode/single' && sP === true) {
+  //     push('/gamemode');
+  //     setsP(false);
+  //     console.log('nal', this.state.hasRead);
+  //   } else if (location.pathname === '/gamemode' && sP === false) {
+  //     push('/gamemode/single');
+  //     setsP(true);
+  //   }
+  // };
 
   //   console.log(history);
+  const clicker = () => {
+    console.log(props);
+  };
   return (
-    (sP && (
-      <Link to="/gamemode">
-        <button onClick={singled}> Single Player</button>
-        <Route path="/gamemode" component={GamemodeButton} />
-      </Link>
-    )) || (
-      <Link to="/gamemode/single">
-        <button onClick={singled}> Single Player</button>
-        <Route path="/gamemode/single" component={GamemodeButton} />
-      </Link>
-    )
+    <div>
+      <button onClick={clicker}>hi</button>
+    </div>
   );
 };
 export default connect(
   state => ({
-    hasRead: state.tasks.hasRead,
+    child: state.child,
+    faceoffs: state.faceoffs,
+    votes: state.votes,
   }),
-  {
-    setTasks: tasks.setTasks.hasRead,
-  }
+  {}
 )(Gamemode);
 // export default connect()(Gamemode);
